requirements.txt
clip==0.2.0
einops==0.8.1
numpy==2.3.1
opencv_python==4.10.0.84
Pillow==11.3.0
qwen_vl_utils==0.0.11
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.4
transformers==4.46.0
